Bug 956899 - Add comments to ConditionVariable and handle some edge cases gracefully; r=froydnj
UltraBlame original commit: cbce330b4dd328a0d7ccc87a1f0d9a1f8ee07a01

+++ js/src/threading/windows/ConditionVariable.cpp
@@ -381,10 +381,12 @@ js::ConditionVariable::wait_for(UniqueLock<Mutex>& lock,
   CRITICAL_SECTION* cs = &lock.lock.platformData()->criticalSection;
 
   
+  
+  
   double msecd = rel_time.ToMilliseconds();
   DWORD msec = msecd < 0.0
                ? 0
-               : msecd > DBL_MAX
+               : msecd > UINT32_MAX
                  ? INFINITE
                  : static_cast<DWORD>(msecd);
